refactor: narrow core materializer and writer ports
